requirements.txt
pytest==7.4.2
selenium==4.13.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
